piece-of-glass: add touchpad-responsive gradient rectangle
This commit is based off of gtk3-demo->gestures from the official gtk
repository. A rectangle that responds to touchpad rotation and zoom
gestures is rendered to the pane.

Compared to upstream, we cache the angle, remove the other gestures,
and preserve the cairo red-violet color gradient (though it'd be
better to change this into user-customizable text or images).